# openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AVAIRA Protocol API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agents/register": {
      "post": {
        "summary": "Register Agent",
        "operationId": "register_agent_api_agents_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents": {
      "get": {
        "summary": "List Agents",
        "operationId": "list_agents_api_agents_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}": {
      "get": {
        "summary": "Get Agent",
        "operationId": "get_agent_api_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/status": {
      "patch": {
        "summary": "Update Agent Status",
        "operationId": "update_agent_status_api_agents__agent_id__status_patch",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/run": {
      "post": {
        "summary": "Agent Run",
        "operationId": "agent_run_api_agents__agent_id__run_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents/{agent_id}/audit": {
      "get": {
        "summary": "Get Agent Audit",
        "operationId": "get_agent_audit_api_agents__agent_id__audit_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "risk_envelope": {
            "$ref": "#/components/schemas/RiskEnvelope",
            "default": {
              "max_spend_usd": 0.0,
              "max_spend_per_action_usd": 0.0,
              "allowed_actions": [],
              "blocked_actions": [],
              "allowed_targets": [],
              "max_concurrent_tasks": 1,
              "require_human_approval_above_usd": 100.0,
              "custom_rules": []
            }
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "goal"
        ],
        "title": "AgentCreate"
      },
      "AgentRunRequest": {
        "properties": {
          "task": {
            "type": "string",
            "title": "Task"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "task"
        ],
        "title": "AgentRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskEnvelope": {
        "properties": {
          "max_spend_usd": {
            "type": "number",
            "title": "Max Spend Usd",
            "default": 0.0
          },
          "max_spend_per_action_usd": {
            "type": "number",
            "title": "Max Spend Per Action Usd",
            "default": 0.0
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions",
            "default": []
          },
          "blocked_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocked Actions",
            "default": []
          },
          "allowed_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Targets",
            "default": []
          },
          "max_concurrent_tasks": {
            "type": "integer",
            "title": "Max Concurrent Tasks",
            "default": 1
          },
          "require_human_approval_above_usd": {
            "type": "number",
            "title": "Require Human Approval Above Usd",
            "default": 100.0
          },
          "custom_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Custom Rules",
            "default": []
          }
        },
        "type": "object",
        "title": "RiskEnvelope"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}